#4 jwt 토큰 검증 개발 :: auth 및 user api 에서 토큰 검증 로직 추가(임시)

+++ src/main/java/com/chatforyou/io/controller/UserController.java
@@ -6,7 +6,9 @@ import com.chatforyou.io.models.in.UserUpdateVo;
 import com.chatforyou.io.models.out.ChatRoomOutVo;
 import com.chatforyou.io.models.out.UserOutVo;
 import com.chatforyou.io.services.AuthService;
+import com.chatforyou.io.services.JwtService;
 import com.chatforyou.io.services.UserService;
+import com.chatforyou.io.services.impl.JwtServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.apache.coyote.BadRequestException;
 import org.springframework.http.HttpStatus;
@@ -23,12 +25,14 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class UserController {
     /**
+     * TODO 개발 필요
      * 이메일 정규식
      * 비밀번호 최소 8자 {영문 특수문자} 포함
      */
 
     private final UserService userService;
     private final AuthService authService;
+    private final JwtService jwtService;
 
     @GetMapping("/info")
     public ResponseEntity<Map<String, Object>> getUserInfo(
@@ -63,7 +67,9 @@ public class UserController {
 
 
     @PatchMapping("/update")
-    public ResponseEntity<Map<String, Object>> updateUser(@RequestBody UserUpdateVo userVo) throws BadRequestException {
+    public ResponseEntity<Map<String, Object>> updateUser(@RequestHeader("Authorization") String bearerToken,
+                                                          @RequestBody UserUpdateVo userVo) throws BadRequestException {
+        jwtService.verifyAccessToken(userVo.getIdx(), bearerToken);
         Map<String, Object> response = new HashMap<>();
         response.put("result", "success");
         response.put("userData", userService.updateUser(userVo));
@@ -71,7 +77,9 @@ public class UserController {
     }
 
     @PatchMapping("/update/pwd")
-    public ResponseEntity<Map<String, Object>> updateUserPasswd(@RequestBody UserUpdateVo userVo) throws BadRequestException {
+    public ResponseEntity<Map<String, Object>> updateUserPasswd(@RequestHeader("Authorization") String bearerToken,
+                                                                @RequestBody UserUpdateVo userVo) throws BadRequestException {
+        jwtService.verifyAccessToken(userVo.getIdx(), bearerToken);
         Map<String, Object> response = new HashMap<>();
         response.put("result", "success");
         response.put("userData", userService.updateUserPwd(userVo));
@@ -79,16 +87,14 @@ public class UserController {
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<Map<String, Object>> deleteUser(@RequestBody UserInVo user) throws BadRequestException {
-        Map<String, Object> response = new HashMap<String, Object>();
-        boolean result = userService.deleteUser(user);
-        if (result) {
-            response.put("result", "success");
-            return new ResponseEntity<>(response, HttpStatus.OK);
-        } else {
-            response.put("result", "fail delete user");
-            return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-        }
+    public ResponseEntity<Map<String, Object>> deleteUser(
+            @RequestHeader("Authorization") String bearerToken,
+            @RequestBody UserInVo user) throws BadRequestException {
+        jwtService.verifyAccessToken(user.getIdx(), bearerToken);
+        userService.deleteUser(user);
+        Map<String, Object> response = new HashMap<>();
+        response.put("result", "success");
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
